refactor(zoom-sdk): split platforms/zoom -> zoom-sdk + zoom-web peers
Part of release 260422-zoom-sdk, Pack F (platform split), slice 1 of 4.
This slice is pure directory rename + path-reference updates; no semantic
change. Imports, Dockerfile COPY paths, binding.gyp, entrypoints, .gitignore,
README, and docs all updated to reference the new paths. History preserved
via git mv — tsc --noEmit has no rename-related errors.

- services/vexa-bot/core/src/platforms/zoom/native/      -> zoom-sdk/native/
- services/vexa-bot/core/src/platforms/zoom/strategies/  -> zoom-sdk/strategies/
- services/vexa-bot/core/src/platforms/zoom/index.ts     -> zoom-sdk/index.ts
- services/vexa-bot/core/src/platforms/zoom/sdk-manager  -> zoom-sdk/sdk-manager.ts
- services/vexa-bot/core/src/platforms/zoom/web/*.ts     -> zoom-web/*.ts
- features/realtime-transcription/zoom/                  -> zoom-sdk/
- features/realtime-transcription/zoom-web/              (new peer, un-gated)

zoom-web relative imports rewritten from ../../../X to ../../X (one
level closer to src/) since zoom-web/ is one level flatter than zoom/web/.

Also includes the approved release artefacts for this cycle:
- tests3/releases/260422-zoom-sdk/groom.md       (9 packs, all approved)
- tests3/releases/260422-zoom-sdk/scope.yaml     (9 issues, summary
                                                   carries PM narrative)
- tests3/releases/260422-zoom-sdk/plan-approval.yaml
                                                  (123 approvals flipped per
                                                   human "approve all" signal)

Slices remaining for Pack F:
- 2: retire ZOOM_WEB env switch; split dispatcher into handleZoomSdk
     + handleZoomWeb as separate entry points
- 3: Platform enum in meeting-api/schemas.py adds ZOOM_SDK + ZOOM_WEB;
     ZOOM becomes deprecated alias -> zoom_sdk
- 4: register 5 static checks for the split (ZOOM_*_PLATFORM_ENUM_EXISTS,
     ZOOM_PLATFORM_DIR_SPLIT, ZOOM_LEGACY_ALIAS_MAPS_TO_SDK,
     ZOOM_SOFT_ENV_SWITCH_RETIRED)

release: 260422-zoom-sdk · stage: develop
