+++ services/vexa-bot/.gitignore
@@ -11,16 +11,16 @@
 *.node
 
 # Zoom SDK (large binaries - download separately)
-# Source code in platforms/zoom/ should NOT be ignored
-core/src/platforms/zoom/native/zoom_meeting_sdk/libmeetingsdk.so*
-core/src/platforms/zoom/native/zoom_meeting_sdk/qt_libs/
-core/src/platforms/zoom/native/zoom_meeting_sdk/*.so*
-core/src/platforms/zoom/native/zoom_meeting_sdk/cpthost
-core/src/platforms/zoom/native/zoom_meeting_sdk/diagnostic/
-core/src/platforms/zoom/native/zoom_meeting_sdk/imjs/
+# Source code in platforms/zoom-sdk/ should NOT be ignored
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/libmeetingsdk.so*
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/qt_libs/
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/*.so*
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/cpthost
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/diagnostic/
+core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/imjs/
 # Keep headers and docs
-!core/src/platforms/zoom/native/zoom_meeting_sdk/h/
-!core/src/platforms/zoom/native/zoom_meeting_sdk/*.md
+!core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/h/
+!core/src/platforms/zoom-sdk/native/zoom_meeting_sdk/*.md
 
 # Environment files
 .env
